Refactor CardFooterChart
